Add doc-blocks for updated classes and new methods
