+++ lib/identity/config.js
@@ -59,16 +59,18 @@ export default class Config {
     const realmName = obj.realmName || obj.realm_name
     const userId = obj.userId || obj.user_id
     const apiUrl = obj.apiUrl || obj.api_url
-    return new this(realmId, realmName, userId, apiUrl)
+    const brokerTargetUrl = obj.brokerTargetUrl || obj.broker_target_url
+    return new this(realmId, realmName, userId, apiUrl, brokerTargetUrl)
   }
 
   /**
    * Create a new instance of Config
    *
-   * @param {string} realmId   The realm's unique identifier
-   * @param {string} realmName The realms globally unique name
-   * @param {string} userId    A specific realm user's unique identifier
-   * @param {string} [apiUrl]  Optional base URL for the Tozny Platform API
+   * @param {string} realmId         The realm's unique identifier
+   * @param {string} realmName       The realms globally unique name
+   * @param {string} userId          A specific realm user's unique identifier
+   * @param {string} [apiUrl]        Optional base URL for the Tozny Platform API
+   * @param {string} brokerTargetUrl The URL which will process broker-based login flows.
    *
    * @returns {Config} The constructed Config object.
    */
