+++ lib/identity/index.js
@@ -32,21 +32,55 @@ import CryptoConsumer from '../utils/cryptoConsumer'
 import { Config as StorageConfig } from '../storage'
 import { PublicKey, SigningKey } from '../types'
 
+/**
+ * Create a identity client connection from the passed identity values.
+ *
+ * @param {Identity} id The Identity instance connected to a specific realm.
+ * @param {string} userId The specific user ID for the user in the configured realm.
+ * @param {StorageConfig} storageConfig The storage client configuration.
+ *
+ * @return {Client} The identity client ready to run identity methods on behalf of a user.
+ */
 function createClient(id, userId, storageConfig) {
   const idConfig = id.config.clone({ userId: userId })
   const storageClient = new id.StorageClient(storageConfig)
   return new Client(idConfig, storageClient, id.crypto)
 }
 
+/**
+ * Identity represents a connection to the Tozny Identity service on behalf of a realm.
+ *
+ * Before registration, login, or other client creation methods are possible, the configuration
+ * for a Tozny Identity realm is needed. Identity holds this configuration and provides methods
+ * for all pre-client operations. In other words, the methods this object make identity clients
+ * for users that belong to the configured realm. It helps authenticate users.
+ */
 export default class Identity extends CryptoConsumer {
+  /**
+   * Check whether a variable is an Identity instance.
+   *
+   * @param {*} identity an object that may or may not be an instance of Identity.
+   *
+   * @return {boolean} Whether or not the passed item is an instance of Identity.
+   */
   static isExtension(identity) {
     return identity.prototype instanceof Identity
   }
 
+  /**
+   * Gets the Client constructor for creating identity Clients.
+   *
+   * @return {Function} The constructor function for creating a Client instance.
+   */
   static get Client() {
     return Client
   }
 
+  /**
+   * Gets the Config constrcutor for creating Identity configuration objects.
+   *
+   * @return {Function} The identity Config constructor function.
+   */
   static get Config() {
     return Config
   }
@@ -62,7 +96,7 @@ export default class Identity extends CryptoConsumer {
    * the storage client constructor in both static _and_ instance method
    * contexts as `this.StorageClient`.
    *
-   * @returns {Client} The storage Client constructor.
+   * @return {Client} The storage Client constructor.
    */
   static get StorageClient() {
     throw new Error(
@@ -89,6 +123,27 @@ export default class Identity extends CryptoConsumer {
     return this.constructor.StorageClient
   }
 
+  /**
+   * Register a new identity with the Tozny identity service.
+   *
+   * This method creates a new identity service user and associated storage identity. Using the
+   * passed username and password it derives encryption keys and writes those credentials to a
+   * Tozny Storage Secure Note. This note is fetchable using the username/password derived keys.
+   *
+   * It also create a broker-based set of recovery notes protected by the provided email that are
+   * used to recover the account in the event the user forgets their password.
+   *
+   * Finally, the fully constructed Client for the user is returned, ready to make requests using
+   * the user identity that was just created.
+   *
+   * @param {string} username The name to associate with the user in the configured realm.
+   *                          may be the same as the email value.
+   * @param {string} password The secret used to protect the users identity and encryption keys.
+   * @param {string} token The registration token to create the storage client with.
+   * @param {string} email The email address used for email brokered access to the identity.
+   *
+   * @return {Client} The identity Client for the user that was just registered with the realm.
+   */
   async register(username, password, token, email) {
     const cryptoKeys = await this.StorageClient.generateKeypair()
     const signingKeys = await this.StorageClient.generateSigningKeypair()
@@ -169,9 +224,9 @@ export default class Identity extends CryptoConsumer {
         max_views: -1,
         expires: false,
         eacp: {
-          email_otp: {
+          email_eacp: {
             email_address: email,
-            template: 'password_recovery',
+            template: 'password_reset',
             provider_link: this.config.brokerTargetUrl
           }
         }
@@ -186,8 +241,8 @@ export default class Identity extends CryptoConsumer {
         max_views: -1,
         expires: false,
         eacp: {
-          note_access: {
-            note_id: brokerKeyNote.noteId
+          last_access_eacp: {
+            last_read_note_id: brokerKeyNote.noteId
           }
         }
       }
@@ -197,6 +252,22 @@ export default class Identity extends CryptoConsumer {
     return idClient
   }
 
+  /**
+   * Get the stored identity credentials for a user and create a Client for them.
+   *
+   * The username and password are used to derive encryption keys used to fetch a pre-stored
+   * note which contains the users identity credentials.
+   *
+   * Broker mode is used when another identity holds the seed (password) used for
+   * the login. Standard login flows with username and password should always use
+   * `broker = false` or omit the parameter.
+   *
+   * @param {string} username The username of the identity to create a Client for.
+   * @param {string} password The secret password of the identity to create a Client for.
+   * @param {boolean} broker Whether or not the password is for a brokered note. False for normal logins.
+   *
+   * @return {Client} The identity Client object for the user.
+   */
   async login(username, password, broker = false) {
     const { noteID, cryptoKeyPair, signingKeyPair } = await deriveNoteCreds(
       this.config,
@@ -222,6 +293,17 @@ export default class Identity extends CryptoConsumer {
     )
   }
 
+  /**
+   * Recreate a identity Client from a serialized representation.
+   *
+   * When storing an identity client, this method will reconstitute the Client from
+   * a serialized representation. The plain JS object is unpacked a new, fully ready
+   * Client instance is returned based on the serialized values.
+   *
+   * @param {Object} obj The serialized Javascript object representing a user.
+   *
+   * @return {Client} The reconstituted identity client for the user.
+   */
   fromObject(obj) {
     if (typeof obj === 'string') {
       try {
@@ -249,14 +331,43 @@ export default class Identity extends CryptoConsumer {
     )
   }
 
+  /**
+   * A wrapper around the broker login flow used for email account recovery.
+   *
+   * @param {string} username The username to recover.
+   * @param {string} recoveryUrl The URL to send the reset initiation to.
+   */
   initiateRecovery(username, recoveryUrl) {
     return this.initiateBrokerLogin(username, recoveryUrl)
   }
 
+  /**
+   * A wrapper around the completion of a broker flow used for email account recovery.
+   *
+   * Once complete a password update should immediately be initiated for the user.
+   *
+   * @param {string} otp The one-time password from the email challenge issued.
+   * @param {string} noteId The ID of the note the email challenge was for.
+   * @param {string} recoveryUrl The URL to send the recovery authentication to.
+   *
+   * @return{Promise<Client>} The recovered identity Client.
+   */
   completeRecovery(otp, noteId, recoveryUrl) {
     return this.completeBrokerLogin(otp, noteId, recoveryUrl)
   }
 
+  /**
+   * Begin a broker-based login flow.
+   *
+   * Broker flows are when another party holds the seed material used to access an
+   * identity account. The broker's access to the seed material is generally protected
+   * by an extra policy check controlled by the user. The initiation request informs
+   * the broker that the user wishes to collect their seed, which causes the broker
+   * to initiate any challenges required by to access the seed material.
+   *
+   * @param {string} username The username of the user wishing to access their credentials.
+   * @param {string} brokerUrl The URL where the broker can be contacted.
+   */
   async initiateBrokerLogin(
     username,
     brokerUrl = `${this.config.apiUrl}/v1/identity/${this.config.realmId}/broker`
@@ -275,6 +386,21 @@ export default class Identity extends CryptoConsumer {
     return validateResponseAsJSON(request)
   }
 
+  /**
+   * Complete a broker based login flow by giving the broker the needed authentication information.
+   *
+   * After initiating a broker-based login, the storage system will issue a challenge to the user,
+   * such as an email containing a one-time password. This information is passed back to the broker
+   * proving the user has granted them access to the seed material. The broker then encrypts and
+   * returns the seed to the user who is then able to derive the keys needed to fetch their identity
+   * credentials anc create a Client instance.
+   *
+   * @param {*} authResponse The authentication material to allow a broker to access the seed material.
+   * @param {string} noteId The ID of the note containing the seed material.
+   * @param {string} brokerUrl The URL where the broker can be contacted.
+   *
+   * @return {Client} An identity Client for the user.
+   */
   async completeBrokerLogin(
     authResponse,
     noteId,
