+++ lib/types/eacp.js
@@ -18,13 +18,22 @@
 
 import Serializable from './serializable'
 import EmailEACP from './emailEACP'
-import NoteAccessEACP from './noteAccessEACP'
+import LastAccessEACP from './lastAccessEACP'
 
 /**
- * NoteOptions represents optional values that are not required for creating a note,
- * but provide additional functionality. Some features are premium and require a TozStore client to work.
+ * EACP defines the various extended access control policies on data object.
+ *
+ * The EACP types are hard coded, and each data object can have one of each
+ * kind. These are defined in sub objects. The EACP object is to mix all of the
+ * various kinds together so they can be sent to the server to apply the policy.
  */
 export default class EACP extends Serializable {
+  /**
+   * Create an EACP instance to organize extended access control policies for data objects.
+   *
+   * @param {EmailEACP} emailEACP An Email EACP configuration to associate with the data object.
+   * @param {LastAccessEACP} noteAccessEACP A Last Access EACP configuration to associate with the object.
+   */
   constructor(emailEACP, noteAccessEACP) {
     super()
 
@@ -36,6 +45,11 @@ export default class EACP extends Serializable {
     }
   }
 
+  /**
+   * Create a plain object representation of the EACP. Used for JSON serialization.
+   *
+   * @return {Object} A plain JS object representing the EACP.
+   */
   serializable() {
     let toSerialize = {}
     // Ensure that plainMeta is always an object, even it it's set to null
@@ -48,14 +62,21 @@ export default class EACP extends Serializable {
     return toSerialize
   }
 
+  /**
+   * Create a new EACP instance from a Javascript object.
+   *
+   * @param {Object} json A plain JS object containing the needed EACP fields.
+   *
+   * @return {EACP} The constructed EACP object based on the passed JS object.
+   */
   static decode(json) {
     let emailEACP
     let noteAccessEACP
     if (typeof json[EmailEACP.jsonKey] === 'object') {
       emailEACP = EmailEACP.decode(json[EmailEACP.jsonKey])
     }
-    if (typeof json[NoteAccessEACP.jsonKey] === 'object') {
-      noteAccessEACP = NoteAccessEACP.decode(json[NoteAccessEACP.jsonKey])
+    if (typeof json[LastAccessEACP.jsonKey] === 'object') {
+      noteAccessEACP = LastAccessEACP.decode(json[LastAccessEACP.jsonKey])
     }
     return new EACP(emailEACP, noteAccessEACP)
   }
