+++ lib/utils/credentials.js
@@ -1,5 +1,14 @@
 import { IDENTITY_DERIVATION_ROUNDS } from './constants'
 
+/**
+ * Derive the note name, crypto, and signing keys for an note containing identity credentials.
+ *
+ * @param {Config} idConfig The identity realm configuration.
+ * @param {Crypto} crypto The concrete Tozny crypto implementation.
+ * @param {string} username The username credentials are being derived for.
+ * @param {*} password The secret password for the user.
+ * @param {*} broker Whether the credentials are for a brokered note.
+ */
 export async function deriveNoteCreds(
   idConfig,
   crypto,
